+++ monitoring/prometheus/alerts.yml
@@ -3,9 +3,17 @@ groups:
     rules:
     - alert: InstanceDown
       expr: up == 0
-      for: 1m
+      for: 10s
       labels:
         severity: page
       annotations:
         description: '{{ $labels.instance }} of job {{ $labels.job }} has been down for more than 1 minute'
         summary: 'Instance {{ $labels.instance }} down'
+    - alert: NetworkLoad
+      expr: histogram_quantile(0.95, sum(rate(ui_request_latency_seconds_bucket[5m])) by (le)) > 0.2
+      for: 10s
+      labels:
+        severity: page
+      annotations:
+        description: '{{ $labels.instance }} of job {{ $labels.job }} has been network high load'
+        summary: 'Instance {{ $labels.instance }} network high load'
