+++ monitoring/prometheus/prometheus.yml
@@ -59,3 +59,11 @@ scrape_configs:
     static_configs:
       - targets:
         - 'post:5000'
+  - job_name: 'docker'
+    static_configs:
+      - targets:
+        - 'docker-host:9323'
+  - job_name: 'telegraf'
+    static_configs:
+      - targets:
+        - 'telegraf:9273'
