+++ monitoring/telegraf/Dockerfile
@@ -1,0 +1,2 @@
+FROM telegraf:1.10.0
+COPY telegraf.conf /etc/telegraf/telegraf.conf
