+++ monitoring/telegraf/telegraf.conf
@@ -1,0 +1,13 @@
+[[inputs.docker]]
+  endpoint = "unix:///var/run/docker.sock"
+  gather_services = false
+  container_names = []
+  container_name_include = []
+  container_name_exclude = []
+  timeout = "5s"
+  perdevice = true
+  total = false
+  docker_label_include = []
+  docker_label_exclude = []
+ [[outputs.prometheus_client]]
+   listen = ":9273"
