--- src/comment_old/VERSION
@@ -1,1 +1,0 @@
-0.0.3
