--- src/comment_old/config.ru
@@ -1,10 +1,0 @@
-require './comment_app'
-require 'rack'
-require 'prometheus/middleware/collector'
-require 'prometheus/middleware/exporter'
-
-use Rack::Deflater, if: ->(_, _, _, body) { body.any? && body[0].length > 512 }
-use Prometheus::Middleware::Collector
-use Prometheus::Middleware::Exporter
-
-run Sinatra::Application
