--- src/post-py_old/requirements.txt
@@ -1,6 +1,0 @@
-prometheus_client==0.0.21
-flask==0.12.3
-pymongo==3.5.1
-structlog==17.2.0
-py-zipkin==0.13.0
-requests==2.18.4
