--- src/ui_old/VERSION
@@ -1,1 +1,0 @@
-0.0.1
