--- src/ui_old/config.ru
@@ -1,14 +1,0 @@
-require 'bundler/setup'
-require 'rack'
-require 'prometheus/middleware/collector'
-require 'prometheus/middleware/exporter'
-require_relative 'ui_app'
-require_relative 'middleware'
-require 'zipkin-tracer'
-
-use Metrics
-use Rack::Deflater, if: ->(_, _, _, body) { body.any? && body[0].length > 512 }
-use Prometheus::Middleware::Collector
-use Prometheus::Middleware::Exporter
-
-run Sinatra::Application
