--- src/ui_old/views/index.haml
@@ -1,27 +1,0 @@
-- unless @posts.nil? or @posts.empty?
-  - @posts.each do |post|
-    %div{ id: 'postlist'}
-      %div{class: 'panel'}
-        %div{class: 'panel-heading'}
-          %div{class: 'text-center'}
-            %div{class: 'row'}
-              .col-sm-1
-                %form{:action => "/post/#{post['_id']['$oid']}/vote/1", :method => "post", id: "form-upvote" }
-                  %input{:type => "hidden", :name => "_method", :value => "post"}
-                  %button{type: "submit", class: "btn btn-default btn-sm"}
-                    %span{ class: "glyphicon glyphicon-menu-up" }
-                %h4{class: 'pull-center'} #{post['votes']}
-                %form{:action => "/post/#{post['_id']['$oid']}/vote/-1", :method => "post", id: "form-downvote"}
-                  %input{:type => "hidden", :name => "_method", :value=> "post"}
-                  %button{type: "submit", class: "btn btn-default btn-sm"}
-                    %span{ class: "glyphicon glyphicon-menu-down" }
-              .col-sm-8
-                %h3{class: 'pull-left'}
-                  %a{href: "/post/#{post['_id']['$oid']}"} #{post['title']}
-              .col-sm-3
-                %h4{class: 'pull-right'}
-                  %small
-                    %em #{Time.at(post['created_at'].to_i).strftime('%d-%m-%Y')}
-                    %br #{Time.at(post['created_at'].to_i).strftime('%H:%M')}
-        .panel-footer
-          %a{href: "#{post['link']}", class: 'btn btn-link'} Go to the link
